refactor: data-model diet — Ore_Kind enum, i32 dijkstra, right-size caps
Audit #15 (M1/P5 remainder): shrink the Game struct's per-cell grids.

- core: replace Ore_Vein.ore_type:string + per-cell color with a single
  Ore_Kind enum byte; add ore_kind_item_id / ore_kind_color to derive the
  content item ID and wall tint from the kind. Ore_Vein drops from 24 B to 1 B.
- core: change Game.dijkstra_map from [N]int to [N]i32 (4 B/cell vs 8 B).
- engine: Engine_Distance_Map backing store is now []i32 (unreachable: i32);
  the get/set API stays int-typed with conversions at the boundary so callers
  are unchanged. Distances are bounded grid hop-counts; i32 is ample.
- gen/gameplay/render: spawn/mine/draw paths use Ore_Kind + derived color/ID.
- io: Save_Ore_Vein is now a single Ore_Kind byte; convert/restore/write drop
  the string+color round-trip. Bump SAVE_VERSION 12 -> 13 (legacy load already
  dropped, single supported version).
- tests: distance-map backing arrays -> i32; save round-trip + layout invariants
  renamed to v13 and assert ore-kind round-trip + Save_Ore_Vein == 1 byte; add
  ore_vein_is_a_single_kind_byte guard; tighten Game-size guard to < 160 KB
  (measured ~132 KB, down from ~222 KB).

+++ src/engine/distance_map.odin
@@ -1,17 +1,21 @@
 package engine
 
+// Distances are small non-negative grid hop-counts (bounded by cell count) plus
+// the sentinel `unreachable`; an i32 backing store halves the per-cell memory of
+// the game's flow-field array versus a machine-word `int`. The get/set API stays
+// `int`-typed so callers do plain integer arithmetic; conversions happen here.
 Engine_Distance_Map :: struct {
-	values:      []int,
+	values:      []i32,
 	grid:        Engine_Grid_2D,
-	unreachable: int,
+	unreachable: i32,
 }
 
 engine_distance_map_make :: proc(
-	values: []int,
+	values: []i32,
 	grid: Engine_Grid_2D,
 	unreachable: int,
 ) -> Engine_Distance_Map {
-	return Engine_Distance_Map{values = values, grid = grid, unreachable = unreachable}
+	return Engine_Distance_Map{values = values, grid = grid, unreachable = i32(unreachable)}
 }
 
 engine_distance_map_is_valid :: proc(dmap: ^Engine_Distance_Map) -> bool {
@@ -33,16 +37,16 @@ engine_distance_map_reset :: proc(dmap: ^Engine_Distance_Map) {
 
 engine_distance_map_get :: proc(dmap: ^Engine_Distance_Map, x, y: int) -> int {
 	if !engine_distance_map_is_valid(dmap) || !engine_grid_2d_contains(dmap.grid, x, y) {
-		return dmap.unreachable if dmap != nil else 0
+		return int(dmap.unreachable) if dmap != nil else 0
 	}
-	return dmap.values[engine_grid_2d_index(dmap.grid, x, y)]
+	return int(dmap.values[engine_grid_2d_index(dmap.grid, x, y)])
 }
 
 engine_distance_map_set :: proc(dmap: ^Engine_Distance_Map, x, y, distance: int) -> bool {
 	if !engine_distance_map_is_valid(dmap) || !engine_grid_2d_contains(dmap.grid, x, y) {
 		return false
 	}
-	dmap.values[engine_grid_2d_index(dmap.grid, x, y)] = distance
+	dmap.values[engine_grid_2d_index(dmap.grid, x, y)] = i32(distance)
 	return true
 }
 
@@ -53,9 +57,9 @@ engine_distance_map_set :: proc(dmap: ^Engine_Distance_Map, x, y, distance: int)
 // and bounds-check (x,y) independently before calling these.
 
 engine_distance_map_get_unchecked :: proc(dmap: ^Engine_Distance_Map, x, y: int) -> int {
-	return dmap.values[engine_grid_2d_index(dmap.grid, x, y)]
+	return int(dmap.values[engine_grid_2d_index(dmap.grid, x, y)])
 }
 
 engine_distance_map_set_unchecked :: proc(dmap: ^Engine_Distance_Map, x, y, distance: int) {
-	dmap.values[engine_grid_2d_index(dmap.grid, x, y)] = distance
+	dmap.values[engine_grid_2d_index(dmap.grid, x, y)] = i32(distance)
 }
